refactor(patina): use ordinary module layout
