+++ crates/vize_patina/src/rules/script/no_deprecated_destroyed_lifecycle/tests.rs
