+++ crates/vize_patina/src/rules/script/prefer_import_meta.rs
@@ -109,5 +109,4 @@ impl PreferImportMetaVisitor<'_> {
 }
 
 #[cfg(test)]
-#[path = "prefer_import_meta_tests.rs"]
 mod tests;
