+++ crates/vize_patina/src/rules/script/tests.rs
