+++ crates/vize_patina/src/rules/vue/no_multiple_template_root/tests.rs
