+++ crates/vize_patina/src/rules/vue/no_unsafe_url.rs
@@ -246,5 +246,4 @@ impl Rule for NoUnsafeUrl {
 }
 
 #[cfg(test)]
-#[path = "no_unsafe_url_tests.rs"]
 mod tests;
