+++ crates/vize_patina/src/rules/vue/no_unsafe_url/tests.rs
