+++ crates/vize_patina/src/rules/vue/permitted_contents.rs
@@ -224,5 +224,4 @@ impl Rule for PermittedContents {
 }
 
 #[cfg(test)]
-#[path = "permitted_contents_tests.rs"]
 mod tests;
