+++ crates/vize_patina/src/rules/vue/permitted_contents/tests.rs
