+++ crates/vize_patina/src/rules/vue/require_toggle_inside_transition/tests.rs
@@ -17,7 +17,9 @@ fn test_invalid_plain_element() {
         result.diagnostics[0].rule_name,
         "vue/require-toggle-inside-transition"
     );
-    insta::assert_debug_snapshot!(result.diagnostics);
+    insta::with_settings!({ snapshot_path => "../snapshots" }, {
+        insta::assert_debug_snapshot!(result.diagnostics);
+    });
 }
 
 #[test]
